refact to rename reservation to booking as the project instructions require
