+++ src/Entity/Booking.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Entity;
+
+class Booking extends Reservation
+{
+}
