+++ src/MessageHandler/BookingCreatedHandler.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace App\MessageHandler;
 
 use App\Message\BookingCreatedMessage;
-use App\Repository\ReservationRepository;
+use App\Repository\BookingRepository;
 use App\Service\MailService;
 use Symfony\Component\Messenger\Attribute\AsMessageHandler;
 use Symfony\Component\Uid\Uuid;
@@ -14,7 +14,7 @@ use Symfony\Component\Uid\Uuid;
 final readonly class BookingCreatedHandler
 {
     public function __construct(
-        private ReservationRepository $reservationRepository,
+        private BookingRepository $bookingRepository,
         private MailService $mailService,
         private \App\Service\NotificationService $notificationService,
     ) {
@@ -22,7 +22,7 @@ final readonly class BookingCreatedHandler
 
     public function __invoke(BookingCreatedMessage $message): void
     {
-        $reservation = $this->reservationRepository->find(Uuid::fromString($message->reservationId));
+        $reservation = $this->bookingRepository->find(Uuid::fromString($message->reservationId));
         if ($reservation === null) {
             return;
         }
